+++ server/src/main/java/org/allaymc/server/command/defaults/OpCommand.java
@@ -2,8 +2,10 @@ package org.allaymc.server.command.defaults;
 
 import org.allaymc.api.command.SimpleCommand;
 import org.allaymc.api.command.tree.CommandTree;
+import org.allaymc.api.entity.interfaces.EntityPlayer;
 import org.allaymc.api.i18n.TrKeys;
-import org.allaymc.api.server.Server;
+
+import java.util.Collection;
 
 /**
  * @author daoge_cmd
@@ -17,24 +19,28 @@ public class OpCommand extends SimpleCommand {
     @Override
     public void prepareCommandTree(CommandTree tree) {
         tree.getRoot()
-                .str("playerName")
+                .playerTarget("player")
                 .exec(context -> {
-                    // TODO: Support offline player
-                    String playerName = context.getResult(0);
-                    var player = Server.getInstance().getPlayerService().getOnlinePlayerByName(playerName);
-                    if (player == null) {
-                        context.addError("%" + TrKeys.M_COMMANDS_GENERIC_PLAYER_NOTFOUND);
+                    Collection<EntityPlayer> players = context.getResult(0);
+                    if (players.isEmpty()) {
+                        context.addPlayerNotFoundError();
                         return context.fail();
                     }
 
+                    if (players.size() > 1) {
+                        context.addTooManyTargetsError();
+                        return context.fail();
+                    }
+
+                    var player = players.stream().findFirst().get();
                     if (player.isOperator()) {
-                        context.addError("%" + TrKeys.M_COMMANDS_OP_FAILED, playerName);
+                        context.addError("%" + TrKeys.M_COMMANDS_OP_FAILED, player.getOriginName());
                         return context.fail();
                     }
 
                     player.setOperator(true);
 
-                    context.addOutput(TrKeys.M_COMMANDS_OP_SUCCESS, playerName);
+                    context.addOutput(TrKeys.M_COMMANDS_OP_SUCCESS, player.getOriginName());
                     player.sendTr(TrKeys.M_COMMANDS_OP_MESSAGE);
                     return context.success();
                 });